chore: Move `get_set` macro into `kv_store` mod
